// src/lib/api/generated/model/presignedUploadInitRequest.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * Etechs Social Network
 * API contract cho social-network-ui (main API + message service).
 * OpenAPI spec version: 1.0
 */
import type { PresignedUploadInitRequestAccess } from './presignedUploadInitRequestAccess';
import type { PresignedUploadInitRequestPurpose } from './presignedUploadInitRequestPurpose';

export interface PresignedUploadInitRequest {
  filename: string;
  content_type: string;
  size_bytes: number;
  checksum_sha256?: string;
  purpose?: PresignedUploadInitRequestPurpose;
  access?: PresignedUploadInitRequestAccess;
}