// src/lib/api/generated/model/presignedUploadInitResponse.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Etechs Social Network
 * API contract cho social-network-ui (main API + message service).
 * OpenAPI spec version: 1.0
 */
import type { NameValueHeader } from './nameValueHeader';
import type { PresignedUploadInitResponseMethod } from './presignedUploadInitResponseMethod';

export interface PresignedUploadInitResponse {
  upload_id: string;
  asset_id: string;
  method: PresignedUploadInitResponseMethod;
  upload_url: string;
  headers: NameValueHeader[];
  expires_at: string;
}