// src/lib/api/generated/model/privacyField.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Etechs Social Network
 * API contract cho social-network-ui (main API + message service).
 * OpenAPI spec version: 1.0
 */

export type PrivacyField = typeof PrivacyField[keyof typeof PrivacyField];


export const PrivacyField = {
  display_name: 'display_name',
  birth_date: 'birth_date',
  bio: 'bio',
  avatar: 'avatar',
  background: 'background',
  personal_info: 'personal_info',
} as const;